Suppress 'deprecated' warnings category
The 'given - when - default' syntax (use in lib/App/perlhl.pm) will be
deprecated in perl-5.38.0 and warnings will be generated whenever
encountered.  The test suite in this distribution depends on matching
test output to expectations.  So the only way, in the short-term, to
have tests continue to pass is to suppress those warnings in the module.